Applicative Population transformer
