+++ src/Control/Applicative/List.hs
@@ -4,11 +4,7 @@ module Control.Applicative.List where
 
 -- base
 import Control.Applicative
--- transformers
-import Control.Monad.Trans.Writer.Strict
 import Data.Functor.Compose
--- log-domain
-import Numeric.Log (Log)
 
 -- * Applicative ListT
 
@@ -17,22 +13,11 @@ import Numeric.Log (Log)
 newtype ListT m a = ListT {getListT :: Compose m [] a}
   deriving newtype (Functor, Applicative, Alternative)
 
+listT :: m [a] -> ListT m a
+listT = ListT . Compose
+
 lift :: (Functor m) => m a -> ListT m a
 lift = ListT . Compose . fmap pure
 
 runListT :: ListT m a -> m [a]
 runListT = getCompose . getListT
-
--- * Applicative Population transformer
-
--- WriterT has to be used instead of WeightedT,
--- since WeightedT uses StateT under the hood,
--- which requires a Monad (ListT m) constraint.
-newtype PopulationT m a = PopulationT {getPopulationT :: WriterT (Log Double) (ListT m) a}
-  deriving newtype (Functor, Applicative, Alternative)
-
-runPopulationT :: PopulationT m a -> m [(a, Log Double)]
-runPopulationT = runListT . runWriterT . getPopulationT
-
-fromWeightedList :: m [(a, Log Double)] -> PopulationT m a
-fromWeightedList = PopulationT . WriterT . ListT . Compose
